requirements.txt
requests
python-dotenv
alpha_vantage
pandas
matplotlib